+++ src/buffer.h
@@ -15,21 +15,21 @@ enum buffer_type {
 
 /** Line metadata */
 struct line {
-    unsigned int len; /** length of line */
+    size_t len; /** length of line */
     char* data;       /** line data */
 };
 
 
 /** A buffer gap is like a cursor in the document */
 struct buffer_gap {
-    unsigned int line;
-    unsigned int col;
-    unsigned int size;
+    size_t line;
+    size_t col;
+    size_t size;
 };
 
 struct buffer {
     /*unique identifier of each buffer for debugging*/
-    unsigned int id;
+    size_t id;
     enum buffer_type type;
     const char* name; /* name of buffer(file) */
 
@@ -37,7 +37,7 @@ struct buffer {
     struct buffer_gap gap;
     /** lines */
     struct line* lines;
-    unsigned int linecount;
+    size_t linecount;
 };
 
 /** Create a new empty buffer of the specified type */
@@ -59,28 +59,28 @@ void buf_delch(struct buffer* b, struct cursor cur);
 void buf_addline(struct buffer* b, struct cursor cur);
 
 /** Return pointer to the line number */
-struct line* buf_line(const struct buffer* buf, unsigned int num);
+struct line* buf_line(const struct buffer* buf, size_t num);
 /** Return the number of newlines '\n' in the buffer */
-unsigned int buf_line_count(const struct buffer* buf);
+size_t buf_line_count(const struct buffer* buf);
 /** Return 1 if line has gap */
-int buf_line_hasgap(const struct buffer* buf, unsigned int line);
+int buf_line_hasgap(const struct buffer* buf, size_t line);
 
 /** Save buffer to disk at path*/
 int buf_save_to_disk(const struct buffer* buf, const char* path);
 
 /** Count number of occurences of char in buffer between cursors
  * from and including 'from' upto and excluding 'to' */
-unsigned int buf_charcount_sec(const struct buffer* b, char ch,
+size_t buf_charcount_sec(const struct buffer* b, char ch,
                                struct cursor from, struct cursor to);
 /** Count number of occurences of char in the whole buffer */
-unsigned int buf_charcount(const struct buffer* buf, char ch);
+size_t buf_charcount(const struct buffer* buf, char ch);
 /** Return char at pos. check buf_in_gap().*/
 char buf_charat(const struct buffer* b, struct cursor cur);
 
-int buf_ingap(const struct buffer* b, unsigned int i);
+int buf_ingap(const struct buffer* b, size_t i);
 
 /** HELPERS */
 
 void buf_pprint(const struct buffer* buf);
-void buf_printline(const struct buffer* buf, unsigned int i);
+void buf_printline(const struct buffer* buf, size_t i);
 void buf_pprint_lines(const struct buffer* buf);
